Implement Linear Sqaures Calculation to reduce LLM error
This one was a doozy. I was able to implement the linear squares algorithm to assist us in straightening out the LLM; it would typically give nutritional perscriptions that were either too high or too low in calories or a given macronutrient. This linear squares solution definetly is not perfect, bu t it is at least a step in the right direction